requirements.txt
tensorflow==1.15.2
Keras==2.2.4
tensorboard==1.14.0
tensorflow-estimator
sklearn
import copy
matplotlib
googledrivedownloader
pandas	
Pillow
google.colab
absl-py
astor
backports.weakref==1.0.post1
enum34
funcsigs
futures
gast
google-pasta
grpcio
h5py
Keras-Applications
Keras-Preprocessing
Markdown
mock
numpy
opencv-python
protobuf
PyYAML
scipy
six
termcolor
Theano
Werkzeug
wrapt